requirements.txt
openai==1.35.12
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
fire==0.6.0